+++ python/examples/spot_cam/ptz.py
@@ -23,7 +23,7 @@ class PtzCommands(Subcommands):
     def __init__(self, subparsers, command_dict):
         super(PtzCommands, self).__init__(subparsers, command_dict, [
             PtzListPtzCommand, PtzGetPtzPositionCommand, PtzGetPtzVelocityCommand,
-            PtzSetPtzPositionCommand, PtzSetPtzVelocityCommand
+            PtzSetPtzPositionCommand, PtzSetPtzVelocityCommand, PtzInitializeLensCommand
         ])
 
 
@@ -119,3 +119,21 @@ class PtzSetPtzVelocityCommand(Command):
             ptz_desc, options.pan, options.tilt, options.zoom)
 
         return ptz_velocity
+
+
+
+class PtzInitializeLensCommand(Command):
+    """Initializes the PTZ autofocus or resets it if already initialized"""
+
+    NAME = 'initialize_lens'
+
+    def __init__(self, subparsers, command_dict):
+        super(PtzInitializeLensCommand, self).__init__(subparsers, command_dict)
+        self._parser.add_argument('ptz_name', default='digi', const='digi', nargs='?', choices=[
+            'digi', 'full_digi', 'mech', 'overlay_digi', 'full_pano', 'overlay_pano'
+        ])
+
+    def _run(self, robot, options):
+        resp = robot.ensure_client(PtzClient.default_service_name).initialize_lens()
+
+        return resp